Corrections after code review
