+++ src/main/kotlin/com/mqgateway/core/hardware/simulated/SimulatedConnector.kt
@@ -1,0 +1,8 @@
+package com.mqgateway.core.hardware.simulated
+
+import com.mqgateway.core.io.provider.HardwareConnector
+
+data class SimulatedConnector @JvmOverloads constructor(
+  val pin: Int,
+  val initialValue: String? = null
+) : HardwareConnector
